Added thread pool capability for publisher
