+++ src/main/java/org/craftercms/studio/impl/v1/service/deployment/job/DeployContentToEnvironmentStore.java
@@ -17,52 +17,24 @@
  */
 package org.craftercms.studio.impl.v1.service.deployment.job;
 
-
-import java.text.SimpleDateFormat;
-import java.time.ZoneOffset;
-import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.locks.ReentrantLock;
 
-import org.apache.commons.lang3.StringUtils;
-import org.craftercms.studio.api.v1.constant.DmConstants;
-import org.craftercms.studio.api.v1.constant.StudioConstants;
-import org.craftercms.studio.api.v1.dal.PublishRequest;
-import org.craftercms.studio.api.v1.dal.SiteFeed;
-import org.craftercms.studio.api.v1.exception.ServiceLayerException;
-import org.craftercms.studio.api.v1.exception.SiteNotFoundException;
 import org.craftercms.studio.api.v1.log.Logger;
 import org.craftercms.studio.api.v1.log.LoggerFactory;
 import org.craftercms.studio.api.v1.repository.ContentRepository;
 import org.craftercms.studio.api.v1.service.GeneralLockService;
 import org.craftercms.studio.api.v1.service.activity.ActivityService;
 import org.craftercms.studio.api.v1.service.configuration.ServicesConfig;
-import org.craftercms.studio.api.v1.service.deployment.DeploymentException;
 import org.craftercms.studio.api.v1.service.deployment.PublishingManager;
 import org.craftercms.studio.api.v1.service.event.EventService;
 import org.craftercms.studio.api.v1.service.site.SiteService;
-import org.craftercms.studio.api.v1.to.DeploymentItemTO;
-import org.craftercms.studio.api.v1.to.PublishingTargetTO;
 import org.craftercms.studio.api.v1.util.StudioConfiguration;
 import org.craftercms.studio.api.v2.service.notification.NotificationService;
 import org.craftercms.studio.impl.v1.job.RepositoryJob;
 import org.springframework.core.task.TaskExecutor;
 
 import static org.craftercms.studio.api.v1.util.StudioConfiguration.JOB_DEPLOYMENT_MASTER_PUBLISHING_NODE;
-import static org.craftercms.studio.api.v1.util.StudioConfiguration.JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_MANDATORY_DEPENDENCIES_CHECK_ENABLED;
-import static org.craftercms.studio.api.v1.util.StudioConfiguration.JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_PROCESSING_CHUNK_SIZE;
-import static org.craftercms.studio.api.v1.util.StudioConfiguration.JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_STATUS_MESSAGE_BUSY;
-import static org.craftercms.studio.api.v1.util.StudioConfiguration.JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_STATUS_MESSAGE_IDLE;
-import static org.craftercms.studio.api.v1.util.StudioConfiguration.JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_STATUS_MESSAGE_PUBLISHING;
-import static org.craftercms.studio.api.v1.util.StudioConfiguration.JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_STATUS_MESSAGE_QUEUED;
-import static org.craftercms.studio.api.v1.util.StudioConfiguration.JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_STATUS_MESSAGE_STOPPED_ERROR;
 
 
 public class DeployContentToEnvironmentStore extends RepositoryJob {
@@ -119,45 +91,10 @@ public class DeployContentToEnvironmentStore extends RepositoryJob {
             Set<String> siteNames = siteService.getAllAvailableSites();
             if (siteNames != null && siteNames.size() > 0) {
                 for (String site : siteNames) {
-                    try {
-                        syncRepository(site);
-                    } catch (Exception e) {
-                        logger.error("Failed to sync database from repository for site " + site, e);
-                        siteService.enablePublishing(site, false);
-                    }
-                    if (siteService.isPublishingEnabled(site)) {
-                        if (!publishingManager.isPublishingBlocked(site)) {
-                            String statusMessage = StringUtils.EMPTY;
-                            try {
-                                Set<String> environments = getAllPublishingEnvironments(site);
-                                for (String environment : environments) {
-                                    logger.debug("Processing content ready for deployment for site \"{0}\"", site);
-                                    List<PublishRequest> itemsToDeploy =
-                                            publishingManager.getItemsReadyForDeployment(site, environment);
 
-                                    if (itemsToDeploy != null && itemsToDeploy.size() > 0) {
-                                        logger.info("Starting publishing on environment " + environment + " for site " + site);
-                                        logger.debug("Site \"{0}\" has {1} items ready for deployment",
-                                                site, itemsToDeploy.size());
-
-                                        PublisherTask publisherTask = new PublisherTask(site, itemsToDeploy,
-                                                environment, studioConfiguration, siteService, publishingManager,
-                                                servicesConfig,contentRepository, activityService, notificationService);
-                                        taskExecutor.execute(publisherTask);
-                                    }
-                                }
-                            } catch (Exception err) {
-                                logger.error("Error while executing deployment to environment store for site: "
-                                        + site, err);
-                                notificationService.notifyDeploymentError(site, err);
-                                logger.info("Continue executing deployment for other sites.");
-                            }
-                        } else {
-                            logger.info("Publishing is blocked for site " + site);
-                        }
-                    } else {
-                        logger.info("Publishing is disabled for site " + site);
-                    }
+                    PublisherTask publisherTask = new PublisherTask(site, studioConfiguration, siteService,
+                            publishingManager, servicesConfig,contentRepository, activityService, notificationService);
+                    taskExecutor.execute(publisherTask);
                 }
             }
         } catch (Exception err) {
@@ -166,54 +103,9 @@ public class DeployContentToEnvironmentStore extends RepositoryJob {
         }
     }
 
-    private Set<String> getAllPublishingEnvironments(String site) {
-        Set<String> environments = new HashSet<String>();
-        if (servicesConfig.isStagingEnvironmentEnabled(site)) {
-            environments.add(servicesConfig.getLiveEnvironment(site));
-            environments.add(servicesConfig.getStagingEnvironment(site));
-        } else {
-            List<PublishingTargetTO> publishingTargets = siteService.getPublishingTargetsForSite(site);
-
-            if (publishingTargets != null && publishingTargets.size() > 0) {
-                for (PublishingTargetTO target : publishingTargets) {
-                    if (StringUtils.isNotEmpty(target.getRepoBranchName())) {
-                        environments.add(target.getRepoBranchName());
-                    }
-                }
-            }
-        }
-        return environments;
-    }
-
-    private void syncRepository(String site) throws SiteNotFoundException {
-        logger.debug("Getting last verified commit for site: " + site);
-        SiteFeed siteFeed = siteService.getSite(site);
-        String lastProcessedCommit = siteFeed.getLastVerifiedGitlogCommitId();
-        if (StringUtils.isNotEmpty(lastProcessedCommit)) {
-            logger.debug("Syncing database with repository for site " + site + " from last processed commit "
-                    + lastProcessedCommit);
-            siteService.syncDatabaseWithRepo(site, lastProcessedCommit);
-        } else {
-            logger.debug("Syncing database with repository for site " + site + " from initial commit");
-            siteService.syncDatabaseWithRepo(site, contentRepository.getRepoFirstCommitId(site));
-        }
-    }
-
     public boolean isMasterPublishingNode() {
         boolean toReturn =
                 Boolean.parseBoolean(studioConfiguration.getProperty(JOB_DEPLOYMENT_MASTER_PUBLISHING_NODE));
-        return toReturn;
-    }
-
-    public boolean isMandatoryDependenciesCheckEnabled() {
-        boolean toReturn = Boolean.parseBoolean(studioConfiguration.getProperty(
-                JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_MANDATORY_DEPENDENCIES_CHECK_ENABLED));
-        return toReturn;
-    }
-
-    public int getProcessingChunkSize() {
-        int toReturn = Integer.parseInt(studioConfiguration.getProperty(
-                JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_PROCESSING_CHUNK_SIZE));
         return toReturn;
     }
