+++ src/main/resources/crafter/studio/studio-services-context.xml
@@ -419,9 +419,11 @@
 
     <bean id="studioPublisherTaskExecutor"
           class="org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor">
-        <property name="corePoolSize" value="10" />
-        <property name="maxPoolSize" value="20" />
-        <property name="threadNamePrefix" value="Studio_Publisher_Thread_Pool" />
+        <property name="corePoolSize"
+                  value="#{studioConfiguration.getProperty(T(org.craftercms.studio.api.v1.util.StudioConfiguration).PUBLISHING_THREAD_POOL_CORE_POOL_SIZE)}" />
+        <property name="maxPoolSize"
+                  value="#{studioConfiguration.getProperty(T(org.craftercms.studio.api.v1.util.StudioConfiguration).PUBLISHING_THREAD_POOL_MAX_POOL_SIZE)}" />
+        <property name="threadNamePrefix" value="#{studioConfiguration.getProperty(T(org.craftercms.studio.api.v1.util.StudioConfiguration).PUBLISHING_THREAD_POOL_NAME_PREFIX)}" />
     </bean>
 
     <bean id="cstudioDeployContentToEnvironmentJobs" class="org.craftercms.studio.impl.v1.service.deployment.job.DeployContentToEnvironmentStore">
